소셜 로그인 구글/AWS-S3 이미지 업로드

+++ src/main/java/com/farmorai/backend/controller/NoticeController.java
@@ -4,13 +4,22 @@ package com.farmorai.backend.controller;
 import com.farmorai.backend.dto.NoticeDto;
 import com.farmorai.backend.dto.PageRequestDto;
 import com.farmorai.backend.dto.PageResponseDto;
+
+import com.farmorai.backend.securityFilter.CustomUserDetails;
 import com.farmorai.backend.service.NoticeService;
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Map;
+
+
+import static org.springframework.http.HttpStatus.*;
+
+
 
 @RestController
 @RequiredArgsConstructor
@@ -30,11 +39,15 @@ public class NoticeController {
     }
 
     @PostMapping
-    public Map<String,String> insertNotice(@RequestBody NoticeDto noticeDto) {
-
-        noticeService.insertNotice(noticeDto);
-        return Map.of("result","success");
+    public ResponseEntity<?> insertNotice(@AuthenticationPrincipal CustomUserDetails userDetails, //로그인한 사용자 정보를 가져옵니다.
+                                           @RequestBody NoticeDto noticeDto) {
+        if(userDetails == null || !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ROLE_ADMIN"))){
+            return ResponseEntity.status(FORBIDDEN).body("관리자만 글 등록이 가능합니다.");
+        }
+        noticeService.insertNotice(userDetails.getMemberId(),noticeDto);
+        return ResponseEntity.ok("글 등록 성공");
     }
+
 
     @PutMapping("/{noticeId}")
     public Map<String,String> updateNotice(@PathVariable Long noticeId, @RequestBody NoticeDto noticeDto) {
@@ -43,8 +56,12 @@ public class NoticeController {
     }
 
     @DeleteMapping("/{noticeId}")
-    public Map<String,String> deleteNotice(@PathVariable Long noticeId) {
+    public ResponseEntity<?> deleteNotice(@AuthenticationPrincipal CustomUserDetails userDetails,
+                                           @PathVariable Long noticeId) {
+        if(userDetails == null || !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ROLE_ADMIN"))){
+            return ResponseEntity.status(FORBIDDEN).body("관리자만 공지사항을 삭제할 수 있습니다.");
+        }
         noticeService.deleteNotice(noticeId);
-        return Map.of("result","success");
+        return ResponseEntity.ok("공지사항이 삭제되었습니다.");
     }
 }
